Ad hoc solution for step size

+++ cpp_hojdoj/engine/include/data_types.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include "Box2D/Box2D.h"
+
 namespace hojdoj{
 
     const double EPSILON = 0.00001;
 
     using Size = unsigned int;
     using Index = unsigned int;
-    using Coord = int;
+    using Coord = float32;
 
     using Time = double;
 }
